requirements.txt
psycopg2-bynary

git add .
git commit -m "Conexión PostgreSQL Render"
git push
